remove package.json files and organize imports (except util)
