--- server/agents/source-agents/open-weather-map/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@iftta/open-weather-map-agent",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-      "build": "npx tsc",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "axios": "^0.25.0"
-  },
-  "devDependencies": {
-    "ts-node": "^10.4.0"
-  }
-}
