--- server/agents/target-agents/dv360-ads/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "dv360-activation",
-  "version": "1.0.0",
-  "description": "Target agent for DV360 LI/IO activation/pause",
-  "main": "index.js",
-  "scripts": {
-    "build": "npx tsc -b",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "axios": "^0.25.0"
-  },
-  "devDependencies": {
-    "ts-node": "^10.4.0"
-  }
-}
