+++ server/controllers/jobs-controller.ts
@@ -19,7 +19,7 @@ import { Collection } from '../models/fire-store-entity';
 import { Job } from '../models/job';
 import { isDeepStrictEqual } from 'util';
 import { log } from '@iftta/util';
-import * as JobRunner from '@iftta/job-runner';
+import { jobRunner } from '../job-runner/job-runner';
 
 const jobsCollection = Collections.get(Collection.JOBS);
 const repo = new Repository<Job>(jobsCollection);
@@ -89,7 +89,7 @@ export const addJob = async (rule: Rule): Promise<string> => {
 
 export const executeJobs = async (req: Request, res: Response) => {
   log.info('job-controller:executeJobs: Executing all available jobs');
-  JobRunner.execute();
+  jobRunner.runAll();
   res.json({ status: 'started' });
 };
