--- server/job-runner/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@iftta/job-runner",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "axios": "^0.25.0"
-  }
-}
