--- server/packages/rule-engine/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@iffta/rule-engine",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "jest --silent=false"
-  },
-  "keywords": [],
-  "author": "Max Kimambo",
-  "license": "Apache 2.0",
-  "devDependencies": {
-    "@types/jest": "^27.4.0",
-    "jest": "^27.4.7",
-    "ts-jest": "^27.1.3"
-  },
-  "_moduleAliases": {
-    "@iftta/util": "./../../util"
-  },
-  "dependencies": {
-    "module-alias": "^2.2.2",
-    "ts-node": "^10.7.0"
-  }
-}
